+++ src/main/java/org/promgrammers/springbootbasic/domain/voucher/service/VoucherService.java
@@ -10,7 +10,7 @@ import org.promgrammers.springbootbasic.domain.voucher.dto.response.VoucherRespo
 import org.promgrammers.springbootbasic.domain.voucher.model.Voucher;
 import org.promgrammers.springbootbasic.domain.voucher.model.VoucherType;
 import org.promgrammers.springbootbasic.domain.voucher.repository.VoucherRepository;
-import org.promgrammers.springbootbasic.exception.BusinessException;
+import org.promgrammers.springbootbasic.global.error.exception.BusinessException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -19,9 +19,9 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.UUID;
 
-import static org.promgrammers.springbootbasic.exception.ErrorCode.DUPLICATED_VOUCHER;
-import static org.promgrammers.springbootbasic.exception.ErrorCode.NOT_FOUND_CUSTOMER;
-import static org.promgrammers.springbootbasic.exception.ErrorCode.NOT_FOUND_VOUCHER;
+import static org.promgrammers.springbootbasic.global.error.exception.ErrorCode.DUPLICATED_VOUCHER;
+import static org.promgrammers.springbootbasic.global.error.exception.ErrorCode.NOT_FOUND_CUSTOMER;
+import static org.promgrammers.springbootbasic.global.error.exception.ErrorCode.NOT_FOUND_VOUCHER;
 
 @Service
 public class VoucherService {
