+++ src/main/java/org/promgrammers/springbootbasic/global/error/exception/DataValidationException.java
@@ -1,4 +1,4 @@
-package org.promgrammers.springbootbasic.exception;
+package org.promgrammers.springbootbasic.global.error.exception;
 
 public class DataValidationException extends RuntimeException {
