+++ src/main/java/org/promgrammers/springbootbasic/global/error/exception/repository/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package org.promgrammers.springbootbasic.exception.repository;
+package org.promgrammers.springbootbasic.global.error.exception.repository;
 
 public class EntityNotFoundException extends RuntimeException {
