+++ src/main/java/org/promgrammers/springbootbasic/global/error/exception/repository/FileWriteException.java
@@ -1,4 +1,4 @@
-package org.promgrammers.springbootbasic.exception.repository;
+package org.promgrammers.springbootbasic.global.error.exception.repository;
 
 public class FileWriteException extends RuntimeException {
