+++ src/test/java/org/promgrammers/springbootbasic/domain/customer/service/CustomerServiceTest.java
@@ -11,7 +11,7 @@ import org.promgrammers.springbootbasic.domain.customer.repository.impl.JdbcCust
 import org.promgrammers.springbootbasic.domain.voucher.model.FixedAmountVoucher;
 import org.promgrammers.springbootbasic.domain.voucher.model.Voucher;
 import org.promgrammers.springbootbasic.domain.voucher.repository.impl.JdbcVoucherRepository;
-import org.promgrammers.springbootbasic.exception.BusinessException;
+import org.promgrammers.springbootbasic.global.error.exception.BusinessException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
